requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.35,<1.0
sqlalchemy>=2.0,<3.0
alembic>=1.14,<2.0
psycopg2-binary>=2.9,<3.0
pydantic-settings>=2.10,<3.0
python-dotenv>=1.0,<2.0
python-multipart>=0.0.5,<1.0
minio>=7.2,<8.0
httpx>=0.27,<1.0
requests>=2.32,<3.0
cachetools>=5.3,<6.0
Pillow>=10.4,<12.0
prisma>=0.15,<1.0
google-genai>=1.19,<2.0
filetype>=1.2,<2.0